% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/socraticswirl-instructor.R
\name{dashboard}
\alias{dashboard}
\title{Open the SocraticSwirl dashboard}
\usage{
dashboard(demo = FALSE)
}
\arguments{
\item{demo}{whether to launch a version with demo data}
}
\description{
Opens the SocraticSwirl instructor dashboard in a browser.
}

