% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/socraticswirl-instructor.R
\name{socratic_swirl_instructor}
\alias{socratic_swirl_instructor}
\title{Log into an instructor account}
\usage{
socratic_swirl_instructor(username, password)
}
\arguments{
\item{username}{account username}

\item{password}{account password}
}
\description{
Log into an instructor account. This has to be done before accessing the
SocraticSwirl dashboard or submitting a new course.
}

