% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/socraticswirl-instructor.R
\name{socratic_swirl_signup}
\alias{socratic_swirl_signup}
\title{sign up with an instructor account}
\usage{
socratic_swirl_signup(username, password, email)
}
\arguments{
\item{username}{desired username}

\item{password}{your password}

\item{email}{e-mail address, which will be sent a confirmation link}
}
\description{
Sign up an account for teaching and uploading SocraticSwirl lessons.
Note that this requires an email confirmation.
}

