% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_course}
\alias{upload_course}
\title{Upload a Swirl course to the SocraticSwirl server}
\usage{
upload_course(directory)
}
\arguments{
\item{directory}{path to directory of course to upload}
}
\description{
Upload a Swirl course to the SocraticSwirl server. Courses can be
created as described here: http://swirlstats.com/instructors.html.
}
\details{
If you already have an existing course with this title, it will
be replaced.
}

